[refactor] Abstract the Xerces Wrapper into its own reusable class
